Throw memory at postgres.  The defaults are pretty light, this changes that.
Notably, work_mem is up to 32MB (from 8), which gives each connection more
room for processing indexes.  And shared_buffers is up to 1/4 of RAM (32GB
on the m5.8xlarge), which may have something to do with the problems we
are encountering.  I've also enabled logging for temp_files (we don't want
those!) and disabled synchrnous write.  This should all result in a faster
database but we'll need to run to confirm.